refactor: split implement_code.py into focused modules (#655)
Split the 1814-line monolith into 7 focused modules under
implementation/ package:
- parsers.py (289 lines): code extraction, validation, diff parsing
- context.py (188 lines): token estimation, file summarization
- prompts.py (293 lines): prompt construction for code generation
- claude_client.py (234 lines): CLI/SDK client, timeouts, progress
- orchestrator.py (562 lines): main loop, retry logic, file validation
- deprecated.py (252 lines): backward-compat batch-mode functions
- __init__.py (110 lines): re-exports for backward compatibility

implement_code.py is now a 48-line re-export shim. All existing
imports continue to work. Test mock patch targets updated to point
at actual module locations.

Unblocks #641, #642, #643, #647 — TDD workflow can now modify
claude_client.py (~234 lines) instead of the full monolith.
